:recycle: Move Vue components into `components` folder
