+++ tslint.json
@@ -1,3 +1,6 @@
 {
-	"extends": "tslint-config-gluons"
+	"extends": "tslint-config-gluons",
+	"rules": {
+		"object-literal-key-quotes": [true, "as-needed"]
+	}
 }
